fix(bootstrap): put in a separate file instead
