+++ lib/RESTClient.js
@@ -10,6 +10,7 @@ const errors = require('arsenal').errors;
 const shuffle = require('arsenal').shuffle;
 
 const conf = require('../config.json');
+const parseBootstrapList = require('./bootstraplist').parseBootstrapList;
 
 function errorMap(mdError) {
     const map = {
@@ -22,26 +23,6 @@ function errorMap(mdError) {
         NotImplemented: 'NotImplemented',
     };
     return map[mdError] ? map[mdError] : mdError;
-}
-
-/**
- * Turns the bootstrap list into an array of host/port objects.
- *
- * @param {String[]} list - the list to process
- *
- * @return {{ host: string, port: number }[]}
- */
-function parseBootstrapList(list) {
-    return list.map(elem => {
-        const tmp = elem.split(':');
-        const host = tmp[0];
-        if (!tmp[1]) {
-            return { host, port: 9000 };
-        }
-        const port = Number.parseInt(tmp[1], 10);
-        assert.ok(!Number.isNaN(port), `port ${tmp[1]} is not a number`);
-        return { host, port };
-    });
 }
 
 class RESTClient {
